사이드바 연동문제 해결
