+++ frontend/src/pages/analysis.js
@@ -44,15 +44,7 @@ function Analysis() {
 
                 <div className={`analysis-container ${show ? 'shrink' : ''}`}>
                     <>
-                        <div className="header-bar">
-                            {!show && (
-                                <>
-                                    <img onClick={handleShow} style={{ cursor: 'pointer' }} src={historyIcon} alt="History" className="history-icon" />
-                                    <img src={editIcon} alt="Edit" className="edit-icon" />
-                                </>
-                            )}
-                            <span className='logo-text' onClick={() => { navigate('/') }} style={{ cursor: 'pointer' }}>VirusDecode</span>
-                        </div>
+                        
                         <Nav variant="tabs" defaultActiveKey="link0" className="justify-content-center">
                             <Nav.Item>
                                 <Nav.Link eventKey="link0" onClick={() => setTab(0)}>Alignment</Nav.Link>
@@ -78,24 +70,6 @@ function Analysis() {
 
 
 
-            <div className={`sidebar ${show ? 'show' : ''}`}>
-                <div className="sidebar-header">
-
-                    <img className="history-icon" src={historyIcon} onClick={handleClose} style={{ cursor: 'pointer' }} />
-                    <img src={editIcon} alt="Edit" className="edit-icon" />
-                </div>
-                <div className="sidebar-body">
-                    <div>Today</div>
-                    <div>Reference1</div>
-                    <div>Reference2</div>
-                    <br />
-                    <div>Previous 7 days</div>
-                    <div>Reference1</div>
-                    <div>Reference2</div>
-                    <div>Reference3</div>
-
-                </div>
-            </div>
